i#6938 migrate: Include sched stats in query and fix related assert (#7060)
schedule_stats_t::get_total_counts() was not including
scheduler-provided stats, as it was doing its own simple aggregation
instead of calling aggregate_results(). We fix that here. That then
triggers the newly added assert from PR #7057 which checks for the
scheduler-provided value being exactly equal meaning there is no data
available. It fires on the schedule_stats_test, which uses a mock stream
which returns -1 for such a stat, so we end up with a negative value. We
update the assert for that condition.

Issue: #6938