i#6662: view tool spacing fix
In REGDEPS public traces the encodings of instructions printed with the `view`
tool are divided in 4 byte words, 2 words on the first line and an
additional 2 on the second line (if necessary).
Instructions that only have 1 word encodings (e.g., a branch) have their
categories printed right after the encodings of their first word and
don't align with the categories of 2 or more words instructions.

For single word instructions we now add spacing equivalent to a second
word.

Issue #6662