i#7050: Remove preempted and faulting instruction from the drmemtraces. (#7058)
Change the implementation to remove preempted and faulting instructions
from drmemtraces. The objective of the change is to remove non-retired
instructions and the corresponding memrefs from drmemtraces.

When an instruction is removed, a new marker
TRACE_MARKER_TYPE_UNCOMPLETED_INSTRUCTION will be added. Its value is
the encoding of the removed instruction up to the size of a pointer.

Renamed handle_kernel_interrupt_and_markers() to
handle_rseq_abort_marker() since the function has been modified to
handle rseq abort.

In oder to remove preempted instructions and memrefs, a new function
preempted_by_kernel_event() is added to look for KERNEL EVENT marker
which may be preceded by memrefs. If a KERNEL EVENT marker is found with
the same PC, the instruction and any following memrefs are removed.

Add unit tests to cover instruction and memref removed caused by a
KERNEL EVENT.

Update offline-legacy-int-offs.templatex,
offline-burst_aarch64_sys.templatex and signal_invariants.c to account
for removed instructions.

Fixes #7050