i#7067 sched cpu limits: Add thread cpu footprint to schedule_stats
Augments schedule_stats to record how many cores each thread appears
on. This is used to compute an overall histogram of cores per thread,
which is a useful metric for studying migrations over time.  This is
done in a multi-workload-aware manner to avoid problems with duplicate
thread ids across workloads.

This adds new pure-virtual functions to the
schedule_stats_t::histogram_t interface, so subclasses will need to be
updated.

Adds some multi-workload support to the testing helpers in
default_memtrace_stream_t and memref_gen.

Adds a unit test.

Issue: #&067