Add decode/encode support for Intel's direct store extensions.
Two new instructions (MOVDIRI and MOVDIR64B) are added. PREFIX_EXT is used for
MOVDIR64B because the enqueue stores will also go there later.

MOVDIR64B is a strange instruction because the encoded destination operand is a
register but it's really an offset relative to the ES segment register. Perhaps
we should represent this as a memory operand in DR, but this patch does the
easy thing.