Add decode/encode support for Intel's direct store extensions.
Two new instructions (MOVDIRI and MOVDIR64B) are added. PREFIX_EXT is used for
MOVDIR64B because the enqueue stores will also go there later.

MOVDIR64B is a strange instruction because the encoded destination operand is a
register but it's really an offset relative to the ES segment register.
Furthermore, the address prefix controls the size of that register operand,
not the data prefix. To handle this a new operand type TYPE_G_ES_VAR_REG_SIZE is
introduced. In the DR IR this is translated to/from a far base disp operand with
DR_SEG_ES set, a null index register, and zero fixed displacement.

For internal processing of the base register for the far base disp a new
operand size OPSZ_addr is introduced. This operand size is the processor word
size, or half that size if the address prefix is present. OPSZ_addr can be
reused in a later commit to replace the users of resolve_addr_size().

There are some additional shenangians in the disassembler to suppress address
size prefixes and suffixes to match the binutils assembly syntax for MOVDIR64B.