i#3544: RV64: Split PC-sensitive IR tests into another test (#7102)
Disassemble result of instructions like jal and auipc on riscv64 is
PC-sensitve, i.e. depends on the current PC value. We used to omit the
check since address of the buffer that we encode instructions to is
unknown before running the test.

This commit splits IR tests of these instructions out and enables regex
match to handle the varying addresses. It's important to keep it small
and standalone to workaround the inefficient regex implementation of
CMake.

Issue: #3544