Add saving and restoring the vtype and vl vector registers for RISC-V vector extension.
This patch was authored at the Computer Architecture and VLSI Laboratory, Institute of Computer Sciense, Foundation of Reasearch and Technology, Hellas.
It emmits the necessary instructions that handle saving and restoring the vl and vtype during the context switches, when using the vector extension of the RISC-V ISA.
This is required, in order to use the use the vector extension correctly.