i#6238 op cat: Change fence opcode category to STATE (#7112)
The x86 OP_*fence opcodes were marked with inaccurate categories in the
original PR #6308. Here we change all 3 to DR_INSTR_CATEGORY_STATE under
the logic that the STATE category is the one most likely to be
interpreted as a barrier by a simulator. (If we add an ATOMIC category
in the future we'll switch to that.)

Issue: #6238