i#7120: Do not add addend for rela jump slot relocation (#7121)
Despite "rela" relocations having an explicit addend value and it being
set to non-0 in a new Debian glibc, the addend is assumed to *not* be
added to the symbol value when relocating on x86_64 and aarch64 (it does
seem to be added on RISCV).
This is not obvious and not well documented; we have to just behave like
existing loaders behave from experimentation/examination. (Yet another
reason to possibly invert the private loader and let the private copy of
ld.so do all the loading and relocating: #5437).

Tested on a machine where nearly every client test in our suite was
crashing after a glibc update: now they pass. Unfortunately it's not
simple to make automated tests for this: we don't have an existing
framework for relocation testing and it would take non-trivial effort to
construct that, beyond the scope of this fix.

Fixes #7120