i#7113 instr decode cache: move module mapper into raw2trace_shared (#7123)
Moves the module mapper and related functionality into raw2trace_shared.

This is in preparation for further changes that include the module
mapper logic in a new library (drmemtrace_instr_decode_cache) that
provides an instr decode cache (instr_decode_cache_t). To avoid having
to pull in the whole drmemtrace_raw2trace implementation into
drmemtrace_instr_decode_cache, which will end up including it in the
invariant checker and everything that depends on, we separate out the
module mapper into raw2trace_shared which is intended for such cases.

This is a pure code move, no changes were made to logic or style.

Issue: #7113