i#6662 public traces: disable read/write counts in invariant_checker
We previously relied on load and store categories of REGDEPS
instructions to indicate the presence of reads and writes that
follows.
Unfortunately, not all instructions that then perform a read or a
write have a corresponding load or a store category, leading to
false positive "Too many reads/writes" invariants firing.

e.g., no store category for this instruction that performs a write
operation:
```
13123537    10416044:        7446 ifetch       4 byte(s) @ 0x0000aaaaefe73500 00000831 0e0d0d04 load [4byte]       %rv11 %rv12 %rv13 -> %rv11
13123537    10416044:        7446                                             0000000f
13123538    10416044:        7446 read         4 byte(s) @ 0x0000115cffc33ba0 by PC 0x0000aaaaefe73500
13123539    10416044:        7446 write        4 byte(s) @ 0x0000115cffc33ba0 by PC 0x0000aaaaefe73500
```

We disable these invariants for REGDEPS traces.

Issue #6662