i#3544 RV64: Preserve vtype and vl vector registers
This patch was authored at the Computer Architecture and VLSI Laboratory, Institute of Computer Sciense, Foundation of Reasearch and Technology, Hellas.

It emits the necessary instructions that handle saving and restoring the vl and vtype during the context switches, when using the vector extension of the RISC-V ISA. This is required, in order to use the vector extension correctly.

Issue: #3544