i#6662 public traces: add instr_t operation_size API
Adds instr_get_operation_size() and instr_set_operation_size()
public APIs for DR_ISA_REGDEPS instructions, which are currently
the only instructions for which operation_size is defined.

We set operation_size of DR_ISA_REGDEPS instructions with no
operands to OPSZ_0 in the convertion to REGDEPS process, so
decoded and converted instructions look the same.
Previously a converted instruction might have had an
operation_size different than OPSZ_0 even with no register
operands, that would then be corrected during encoding and
subsequent decoding. This is mostly for convenience when testing.

We update a test to also invoke the instr_length() API on
DR_ISA_REGDEPS instructions.

Issue #6662