i#7157 syscall sched: Handle static injected syscall traces in scheduler
Adds handling for statically injected kernel syscall traces in the scheduler.

Ensures that quantum context switches are not done in the middle of a statically-injected syscall trace.

Also ensures that voluntary context switches are also delayed until after the syscall trace. This involved fixing the bookkeeping logic which is done on the next user-space instr.

We keep status quo on the scheduler behavior of showing the post-syscall markers before the switch.

Adds a unit test for statically injected kernel syscall trace handling by the scheduler.

Issue: #7157