i#3544 RV64: Preserve vtype and vl vector registers (#7110)
This patch was authored at the Computer Architecture and VLSI
Laboratory, Institute of Computer Science, Foundation of Research and
Technology, Hellas.

It emits the necessary instructions that handle saving and restoring the
vl and vtype during the context switches, when using the vector
extension of the RISC-V ISA. This is required, in order to use the
vector extension correctly.

Also corrected the order of append_restore_xflags and
append_restore_simd_reg (in emit_utils_shared.c ) in order to have the
restoration happen in the reverse order of the saving. This specific
code is not currently used for RISC-V, but it might in the future.

Issue: #3544