i#7192 module map race: Skip unnecessary unmap in analyzer
Skips the unnecessary unmap before the subsequent map to the same region. There is
potential for a race with other threads that may mmap some memory between the
unmap and map, which will then get stolen because our map uses MAP_FIXED.

Adds TODO to do the same for non analyzer cases.

Issue: #7192