i#3699 ARM: Add padding to priv_mcontext_t for 8-byte alignment. (#7191)
ef4482e13 caused a few additional failures. In particular, the child of
a clone3 system call started with the register values shifted: R2 got
the value that should be in R1 and so on. That caused the test
linux.clone to fail.

It turns out that some callers of insert_{push,pop}_all_registers expect
that an unpadded priv_mcontext_t will be pushed/popped. Moving the
4-byte padding to above the struct instead of below it makes some tests
pass but breaks others. So we add some padding in the middle of
priv_mcontext_t, changing its size from 324 to 328. It will probably run
slightly faster with the struct (and particularly the field "simd")
8-byte aligned.

Update insert_{push,pop}_all_registers and arm.asm for the new struct
layout, with some additional changes to arm.asm:

+ Delete irrelevant X64 macros.

+ Document the macros more clearly (they can be found with grep).

+ Change SP only at start and end of function (standard practice, makes
debugging easier).

+ Avoid writing below SP (compatible with signal handlers).

This change breaks compatibility so the version is increased to 11.90.

Issue: #3699