i#1973: Fix linux.signal* tests on musl
POSIX.1 requires SS_ONSTACK to appear only in a stack_t retrieved from
sigaltstack(), not one passed to it. musl checks the invalid flag and
will bail out with EINVAL before issuing a syscall in this case.

This commit sets ss_flags to zero in the test instead, on both glibc and
musl this should issue a syscall and result in an EPERM failure.

Issue: #1973