i#7199: make adding a new register less error-prone
It is currently easy to miss an important step when adding
a new register, as the value of the last register is marked
by two alias enums that have to be maintained.

Add a new enum `DR_REG_AFTER_LAST_VALID_ENUM` to mark the
end of the actual registers. This means we can avoid using
aliases.

Fixes: #7199