Update dstack_OFFSET.
Since unprotected_context_t has priv_mcontext_t as a field,
increasing the alignment of priv_mcontext_t increases the
alignment of unprotected_context_t, which causes the size
of unprotected_context_t to be increased to make it a multiple
of the alignment, and that increases the offset of dstack in
dcontext_t, because dcontext_t starts with an
unprotected_context_t.

We should really have a
static_assert(dstack_OFFSET == offsetof(dcontext_t, dstack))
instead of just optimistically stating it in a comment!

Change-Id: I201cbc38f94693f2da4d739c2d8886e3c0f64c4c