i#7195 Fix Android library base address
Change the build scripts to use lld's `--image-base=` argument to set the
preferred base address for libraries instead of `-Ttext=`.

`-Ttext=` sets the address for the `.text` section but `lld`'s section
layout does not put the `.text` section first so not all sections are
moved to the preferred base.

Changing it to use `--image-base=` instead means we can remove the
Android special-casing and use the same preferred base address that we
do for Linux.

Fixes: #7195