i#2154 Android64: Deploy c++ runtime library
Android splits the C++ STL runtime into a separate library which is not
provided as a standard system library. Applications are expected to
statically link the STL library (libc++_static.a) or distribute the
dynamic library with the application.

The NDK C++ support guide recommends using the shared library C++
runtime if you are building a shared library or an application that uses
shared libraries [1] so I have set the DynamoRIO CMake scripts to do
that by default and added code to deal with deploying libc++_shared.so.

This is handled in two places:
 - drsyms CMakeLists.txt explicitly installs libc++_shared.so into
   ${INSTALL_EXT_LIB} with the other dependency libraries. This will
   ensure the library is included when generating a DynamoRIO package.

 - DynamoRIO_copy_target_to_device() now checks if a target it is
   copying is a C++ target and automatically copies the runtime library
   as well if it is.

[1] https://developer.android.com/ndk/guides/cpp-support#important_considerations

Issue: #2154, #1874