i#7224 skipped header: Get filetype from stream in opcode_mix
Fixes the opcode_mix tool to get the filetype from the memtrace_stream_t interface instead
of the TRACE_MARKER_TYPE_FILETYPE marker memref which may not be seen if -skip_instrs is
applied by the user.

Adds a new unit test that runs the opcode_mix tool with -skip_instrs, which would fail
because of an unseen filetype prior to this fix.

Issue: #7224, #7113