i#7224 skip docs: Clarify drmemtrace skip option docs
Clarifies that the -{skip,sim,warmup}_refs options skip a count of
trace records inside particular tools, rather than in the framework.

Clarifies that the -skip_{instrs,to_timestamp} options skip over
top-level markers, but their values are available from the stream API.

Adds i#7230 TODO comments on fixing the -{sim,warmup}_refs options in
the simulators to count marker records (or possibly to keep as is and
update the docs).

Issue: #7224, #7230