i#7224 skipped header: Get filetype from stream in opcode_mix (#7225)
Fixes the opcode_mix tool to get the filetype from the memtrace_stream_t
interface instead of the TRACE_MARKER_TYPE_FILETYPE marker memref which
may not be seen if -skip_instrs is applied by the user.

Modifies opcode_mix_t to use initialize_stream instead of initialize, to
get the serial stream.

Adds a new unit test that runs the opcode_mix tool with -skip_instrs,
which would fail because of an unseen filetype prior to this fix.

Augments decode_cache_t::init() documentation to suggest a reliable way
to obtain the filetype.

Issue: #7224, #7113