i#7113 decode cache: Add build arch check
Moves the check that compares the current build enviroment with the arch bit of the trace file to decode_cache_t::init.

Due to some complexities in how build_target_arch_type() is declared, it had to be wrapped by a new API in
decode_cache_base_t. build_target_arch_type() is defined conditionally in trace_entry.h only if dr_defines.h
has already defined certain symbols. It is hard to ensure that for all builds, so we ensure it just for
decode_cache.cpp which is packaged as a separate lib unit drmemtrace_decode_cache.

Issue: #7113