i#7113 decode cache: Refactor view_t to use the new lib (#7219)
Refactors the view_t tool to use the new decode_cache_t instead of
duplicating the instr decoding logic.

Modifies decode_info_base_t::set_decode_info_derived to allow it to
return the error string. This is useful for cases where there was an
error during its operation, which is especially possible when used with
a decode_cache_t with include_decoded_instr_ = false in which case the
decode_info_t also performs the decoding on its own.

Adds new tests to decode_cache_test to verify some related cases. Not
adding a new release note for the API signature change as there was no
intervening release since it was initially added.

Also fixes the view_t logic to get filetype, which should be obtained
from the memtrace_stream_t object instead of TRACE_MARKER_TYPE_FILETYPE
to handle cases where the filetype marker is not seen by the tool
because of analyzer options like -skip_instrs that skip over a certain
initial part of the trace.

Issue: #7113, #7224