i#7113 decode cache: Add build arch check (#7233)
Moves the check that compares the current build enviroment with the arch
bit of the trace file to decode_cache_t::init.

Due to some complexities in how build_target_arch_type() is declared in
trace_entry.h, it had to be wrapped by a new API in decode_cache_base_t.
build_target_arch_type() is defined conditionally in trace_entry.h only
if dr_defines.h has already defined certain symbols. It is hard to
ensure that for all builds dr_defines.h is included before
trace_entry.h, so we ensure it just for decode_cache.cpp which is
packaged as a separate lib unit drmemtrace_decode_cache. This also means
that the new API could not be in decode_cache_t which is a template
class that has to be defined in the header.

Issue: #7113, #7236