i#4800,i#5873: Disable longer test suite on merge
Disables running longer sets of tests on merges to master: both
TEST_LONG for the general test suite and
SKIP_LESS_DIFFERENTIATED_TESTS_FOR_PULL_REQUEST being disabled for
32-bit Windows.

We ideally would run our full set of tests on merges to the master
branch, but we have not had the resources to keep that suite green.
And now that we are turning off the "Require branches to be up to date
before merging" requirement we really want pre and post merge suites
to be identical for ease of monitoring concurrent merge problems.

Issue: #4800, #5873