i#3544: RV64: Avoid UB with varadic arguments in test linux.mmap
When printing the parameters, the fd -1 is interpreted as int, but
printed as long int ("%lx") on 64-bit platforms.

This doesn't cause any issue on x86_64 and aarch64, where passing a
32-bit integral argument in GPR zeroes the high 32 bits, resulting
0xffffffff. But riscv64 ABI requires the argument to be passed after
sign-extension, failing the test with 0xffffffffffffffff printed.

Let's print these 32-bit values precisely with 32-bit flag characters,
avoiding the undefined behaviour.

Issue: #3544