i#7260: tool.drcacheoff.purestatic bug fix
`tool.drcacheoff.purestatic` fails to build due to `libxxhash` missing.
This library is used by `liblz4` in some cases, and must be statically
linked when building the `tool.drcacheoff.purestatic` app, which we use
in some tests.

We update the doc adding `libxxhash-dev` to the list of packages that
need to be installed to build DynamoRIO on Linux.

Note: `tool.drcacheoff.purestatic` is currently only built in x64 Linux
hosts.

Fixes #7260