i#5873: Explicitly request action for failing master merge tests
As we've now removed the requirement for a branch to be up to date before merging to master,
there is a possibility of unintended regressions due to interactions between commits. Such
regressions can be detected by the post-merge test suite which serializes the commits.

Adds an explicit request to triage and fix the possible regression on master merge test
workflow failures.

Issue: #5873