i#7281: Move failure notification config to reusable workflow (#7282)
Refactors the Github Action configs to move logic to send notification
email on master merge failure into a reusable workflow which is now used
by all workflows instead of duplicating the logic 24 times.

Also, each workflow file now has a separate job for the new reusable
workflow that is invoked after all other jobs (that actually run the
test suite) are done. Status of all the test suite jobs are sent as part
of the single failure notification email, if there's at least one failure;
we skip filtering of status message by test suite status mostly for
simplicity of logic that must be duplicated in each workflow file.

Note that the reusable workflow cannot be invoked as part of some step
of a job; it has to be a separate job on its own.

The failure notification job requires a separate worker. Note that since
it is invoked only when all other jobs in the same workflow are done,
this doesn't make our concurrent job quota usage for a single master
merge event any worse. However, as we're still adding a new job, it
indeed uses some quota that may have been used by other concurrent
triggers (PRs or master merges); but note that since it is gated by the
failure condition and is very short anyway, this is not a concern.

Test run:

Email notification:
https://groups.google.com/g/dynamorio-devs/c/EWkaTPBrjww
Notification job in failing workflow:
https://github.com/DynamoRIO/dynamorio/actions/runs/13351946734/job/37289912108?pr=7282
(other notification jobs in non-failing workflows were skipped)

Issue: #7281