i#5873: Explicitly request action for failing master merge tests (#7279)
As we've now removed the requirement for a branch to be up to date
before merging to master, there is a possibility of surprise regressions
on the master branch due to interactions between commits. Such
regressions can be detected by the post-merge test suite which is run on
a serialized view of the commits.

Adds an explicit request to triage and fix the possible regression on
the dynamorio-devs email notification sent out on master merge test
workflow failures.

Issue: #5873