i#5873: Include hash in post-commit CI failure email subject (#7285)
Includes the commit hash in the subject line of notification emails sent
out on post-commit master merge CI failures.

Without this, failures in the same CI workflow at different commits are
merged in the same subject thread. It is better to have them on
different threads as the reasons for the failures could very well be
different.

Issue: #5873