i#7284: Bootstrap Alpine build environment from minirootfs in workflow
Previously, the build environment is bootstrapped with a static version
of apk downloaded from Alpine mirror, whose URL isn't stable since the
package is upgraded irregularly.

This commit switches to prebuilt minirootfs tarball, which is considered
as release artifacts and has a stable URL, saving the burdern of
frequently synchronizing with Alpine upsteram.

Fixes: #7284