i#7157 dyn inject: Improve ord tracking for synthetic records
Fixes handling of some cases related to ordinal tracking when there are queued synthetic records.

get_input_record_ordinal, which has existing logic to discount records that are read-ahead by the scheduler, was missing
logic to correctly account for synthetic records in the queue. This caused a higher count to be subtracted from the
record ordinal returned by the input stream, particularly when a new input is switched to (in which case the guarding
condition for the read-ahead adjustment was satisfied).

Added tracking for whether each queued record is real, that is, read from the input reader or synthetic.

Fixes an issue with instrs_pre_read  tracking which should be decremented only when a real record is passed on to the caller.

Adds a test that verifies the bookkeeping for use_input_ordinals case, which was affected by the above issue.

Issue: #7157