i#7157 dyn inject: Fix extra switch injections
Fixes a bug in the dynamic injection logic in the drmemtrace scheduler that caused the context switch sequence to be injected more times than actual context switches.  It was erroneously injected also when set_cur_input was invoked without any change to the input.

update_switch_stats() is a better control point for switch sequence injection, which is renamed to on_context_switch here. Now, in the switch insertion offline test, we have 11 context switch injections (== input-to-input + idle-to-input switches), instead of the prior 363

Issue: #7157