i#7315: Avoid ID_AA64MMFR2_EL1 under QEMU to avoid fatal SIGILL (#7316)
Under QEMU, proc_init()'s MRS of ID_AA64MMFR2_EL1 causes a fatal SIGILL.
We avoid it when under QEMU.
(Note that this did not fail on Ubuntu20's QEMU, only on Ubuntu22, but
we avoid under any QEMU for now to unblock progress.)

We'd prefer to use TRY_EXCEPT_ALLOW_NO_DCONTEXT, but proc_init() is
called prior to init-time signal handling being set up: and we'd need
to add SIGILL to the ones caught at init time, which complicates later
uses of SIGILL for NUDGESIG_SIGNUM and suspend_signum (and on x86
XSTATE_QUERY_SIG): so we'd want SIGILL to only work for try-except at
init time. That is all a little too involved to implement right now so
we're
putting in this disabling to unblock Ubuntu22 progress for #7270.
Long-term we probably want to put in the try-except, so we'll leave
#7315 open.

Tested on a local Ubuntu22 x86 machine with aarch64 cross-compilation
where every test failed before this fix and nearly all pass now (the
others fail for other reasons masked by this SIGILL before).

Issue: #7270, #7315