i#1569 AArch64: Strip memory tags in vmareas/memquery
Android 11+ uses the scudo heap allocator [1] which always returns
tagged pointers if running on hardware that supports AArch64 TBI (Top
Byte Ignore) or MTE (Memory Tagging Extension). App pointers that were
obtained from malloc or new will contain a tag in the top byte of the
pointer. In most situations this doesn't matter to DynamoRIO, its just
part of the pointer value.

However, memory tags are assigned in userspace by the allocator. They
are not stored in the process's page tables (they would be too granular
for that) and addresses in the process's memory map do not contain tags.
When DynamoRIO looks up an `app_pc` in vmareas or memquery we need to
strip the tag out so it can be compared to the untagged VAs in the
memory map.

This was discovered when testing on AArch64 Android but the same problem
happens with Linux systems using a tagging allocator. I have reproduced
it on Linux with glibc with the env var
`GLIBC_TUNABLES=glibc.mem.tagging=3` set [2].

This fixes asserts seen running the test linux.bad-signal-stack:

dynamorio/core/unix/memcache.c:382 found

CURIOSITY : !exists || vmvector_overlap(all_memory_areas, start, end) ||
    are_dynamo_vm_areas_stale() || !dynamo_initialized in file
    dynamorio/core/unix/memcache.c line 340

[1] https://source.android.com/docs/security/test/scudo
[2] https://www.gnu.org/software/libc/manual/html_node/Memory-Related-Tunables.html

Issue: #1569, #2154