i#3544 RV64: Fix bugs on saving vector registers and multithreaded icache flushing. (#7298)
There was an error in the code of core/arch/riscv64/mangle.c, causing
the vector registers to be saved in wrong addresses.

Also the SYS_RISCV_FLUSH_ICACHE_LOCAL flag was passed to the
riscv_flush_icache system call, which is incorrect.
It causes to flush the instruction cache of only the calling thread,
when multiple threads are used.

This patch fixes both bugs mentioned above.

Issue: #3544