i#7270: sample.callstack output mismatch fix
The `code_api|sample.callstack` test fails to match the expected output
on both x86_64 and aarch64 with Ubuntu 22.04.
This test is supposed to use libcallstack to track malloc calls, but
the binaries we run the test on, which come from eflags.c and nzcv.c,
don't call malloc anymore. Likely, there was an indirect call to malloc
with Ubuntu 20, but not with Ubuntu 22.
To fix this test we directly add a call to malloc in eflags.c and
nzcv.c.

Issue: #7270