i#7340 Ubuntu22: Treat 32-bit AMD syscall as sysenter (#7348)
Maps the 32-bit AMD syscall instruction to TRACE_TYPE_INSTR_SYSENTER. It
is fair to treat this as the sysenter instruction, and the invariant
checker and other tools can now do the same because of
TRACE_TYPE_INSTR_SYSENTER.

As noted in the Linux implementation
(https://github.com/torvalds/linux/blob/4d872d51bc9d7b899c1f61534e3dbde72613f627/arch/x86/entry/entry_64_compat.S#L142),
x86-32 can be assumed to use OP_syscall only on the 32-bit AMD vdso
__kernel_vsyscall. This OP_syscall really acts like an OP_sysenter
though, as the kernel redirects control flow to a hardcoded address in
the __kernel_vsyscall sequence.

Without this, we see a failure in the tool.drcacheoff.windows-invar test
on the x86-32 suite running on Github Actions runners which are observed
to be AMD. Interestingly, this shows up only on Ubuntu22, but not when
the AMD related issues were initially observed
(https://github.com/DynamoRIO/dynamorio/issues/6417).

Tested manually that the test succeeds on a Github Actions runner where
it failed before.

Issue: #7270
Fixes: #7340